feature/WRN-fetch-apiservice: refactor in admin view introducing members component
